Create Card_game.py
In this card game it shuffles the player which goes first. After that 3 different classes such as Card, Player and Deck are used. 
It declare the winner at the end if one of the player runs out of card, the player with cards in hand is a winner!!
Goes at war if one of the player has less then specified length of card  and looses the game!